+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - added `/load` and `/save` commands for even quicker savegame operations
 - added `/demo` command to quickly play the demo
 - added `/title` command to quickly exit to title
+- added `/vsync on` and `/vsync off` commands to toggle the VSync setting
 - added `/give all` variant of the item cheat
 - changed injection files to be placed in its own directory (#1306)
 - changed item cheat sound effects